requirements.txt
grpcio           
grpcio-tools     
psycopg2-binary  
python-dotenv
